feat(alertInLoginPage): remove of the App.vue
