+++ src/stores/user.ts
@@ -6,7 +6,7 @@ import router from "@/router";
 import { useMaterialIconsStore } from "./material-icons";
 import { useDataStore } from "./data";
 import { saveAlertPopup } from "@/components/SaveAlertPopup.vue";
-import { unauthorizedAlertPopup } from "@/utilities/functions";
+import { unauthorizedAlertPopup } from "@/utilities/unauthorizedAlertPopup";
 
 interface UserState {
   user: { email: string };
